dashboard/requirement.txt
keras==2.13.1
numpy==1.24.3
tensorflow==2.13.0
pillow==9.5.0
streamlit==1.25.0
opencv-python==4.6.0.66